update for front end

+++ routes/api.php
@@ -25,10 +25,10 @@ Route::group(['prefix' => 'auth'], function ()
     Route::post('signup', 'AuthController@signup');
 
     Route::post('login', 'AuthController@login');
-    Route::post('logout', 'AuthController@logout');
+    Route::post('logout', 'AuthController@logout')->middleware('auth.role:admin, user');
     Route::post('refresh', 'AuthController@refresh');
-    Route::post('me', 'AuthController@me');
-    Route::post('payload', 'AuthController@payload');
+    Route::post('me', 'AuthController@me');             // ar reikalingas ??
+    Route::post('payload', 'AuthController@payload');   // ar reikalingas ??
 });
 
 
@@ -37,9 +37,9 @@ Route::group([], function()
 {
     Route::get('ads', 'AdsController@index');
     Route::get('ads/{ad}', 'AdsController@show');
-    Route::post('ads', 'AdsController@store');
-    Route::put('ads/{ad}', 'AdsController@update');
-    Route::delete('ads/{ad}', 'AdsController@destroy');
+    Route::post('ads', 'AdsController@store')->middleware('auth.role:admin, user');
+    Route::put('ads/{ad}', 'AdsController@update')->middleware('auth.role:admin, user');
+    Route::delete('ads/{ad}', 'AdsController@destroy')->middleware('auth.role:admin, user');
 }); 
 
 // comment routes
@@ -47,27 +47,27 @@ Route::group([], function()
 {
     Route::get('ads/{ad}/comments', 'CommentsController@index');
     Route::get('ads/{ad}/comments/{comment}', 'CommentsController@show');
-    Route::post('ads/{ad}/comments', 'CommentsController@store');
-    Route::put('ads/{ad}/comments/{comment}', 'CommentsController@update');
-    Route::delete('ads/{ad}/comments/{comment}', 'CommentsController@destroy');
+    Route::post('ads/{ad}/comments', 'CommentsController@store')->middleware('auth.role:admin, user');
+    Route::put('ads/{ad}/comments/{comment}', 'CommentsController@update')->middleware('auth.role:admin, user');
+    Route::delete('ads/{ad}/comments/{comment}', 'CommentsController@destroy')->middleware('auth.role:admin, user');
 });
 // order routes
 Route::group([], function() 
 {
-    Route::get('orders', 'OrdersController@index');
-    Route::get('orders/{order}', 'OrdersController@show');
-    Route::post('orders', 'OrdersController@store');
-    Route::put('orders/{order}', 'OrdersController@update');
-    Route::delete('orders/{order}', 'OrdersController@destroy');
+    Route::get('orders', 'OrdersController@index')->middleware('auth.role:admin');
+    Route::get('orders/{order}', 'OrdersController@show')->middleware('auth.role:admin, user');
+    Route::post('orders', 'OrdersController@store')->middleware('auth.role:user');
+    Route::put('orders/{order}', 'OrdersController@update')->middleware('auth.role:admin, user');
+    Route::delete('orders/{order}', 'OrdersController@destroy')->middleware('auth.role:admin');
 });
 // user routes
 Route::group([], function() 
 {
-    Route::get('users', 'UsersController@index');
-    Route::get('users/{user}', 'UsersController@show');
+    Route::get('users', 'UsersController@index')->middleware('auth.role:admin');
+    Route::get('users/{user}', 'UsersController@show')->middleware('auth.role:admin, user');
     //Route::post('users', 'UsersController@store');
-    Route::put('users/{user}', 'UsersController@update');
-    Route::delete('users/{users}', 'UsersController@destroy');
+    Route::put('users/{user}', 'UsersController@update')->middleware('auth.role:admin, user');
+    Route::delete('users/{users}', 'UsersController@destroy')->middleware('auth.role:admin, user');
 });
